requirements.txt
pinecone-client
openai 
tiktoken
langchain
flask
python-dotenv